python-code/requirements.txt
fastapi
firebase_admin
fuzzywuzzy
gensim
google.generativeai
joblib
numpy
openai
nltk
pandas
pdfplumber
protobuf
pydantic
PyPDF2
python_docx
python-multipart
spacy
scikit_learn
uvicorn
